Restrict combat unit training in Stone Age
In the Stone Age when there are not many build options in a city the AI will spam combat units with the effect that the research slider has to be reduced to produce the gold to maintain them. There is no great advantage at this point to having so many units so the number should be limited by the player population and number of cities.

AC4-290