Remove financial penalties from running Chiefdom and Tribalism civics
These civic is the first new Government and Society options available and comes at a time when there is little opportunity for gold creation. The increased maintenance that was present before this commit had the effect of either preventing a player founding new cities due to the costs or killing research as the player needed to significantly increase Gold slider at the expense of Research.

AC4-236